+++ src/connect/mod.rs
@@ -9,4 +9,6 @@
 //! caching, and health check; and also prevents DDOS and reports metrics 
 //! concerning health status of backend servers.
 //! 
-// ------------------ bring external libraries/crates -------------------------
+// ----------------------------- bring Modules ---------------------------------
+pub mod connection;
+pub mod appargs;
